man idk i did a lot but it works so idk

+++ LevyMarti-UF5-Project/src/DAO/ChampionDAO.java
@@ -43,51 +43,73 @@ public class ChampionDAO extends BaseDAO {
     }
 
     public int insertChampion(Champion item) throws SQLException {
+        TypeDAO tpDao = new TypeDAO();
+        int affectedRows = 0;
+
         String query = "INSERT INTO champion (code, name, shortDesc, winrate, releaseDate, isRanged, price, typeID) VALUES (?,?,?,?,?,?,?,?)";
         PreparedStatement stmt = conn.prepareStatement(query);
 
-        stmt.setInt(1, item.getCode());
-        stmt.setString(2, item.getName());
-        stmt.setString(3, item.getShortDesc());
-        stmt.setInt(4, item.getWinrate());
-        stmt.setDate(5, Date.valueOf(item.getReleaseDate()));
-        stmt.setBoolean(6, item.isRanged());
-        stmt.setDouble(7, item.getPrice());
-        stmt.setInt(8, item.getType().getId());
-
-        int affectedRows = stmt.executeUpdate();
+        if (tpDao.getType(item.getType().getId()) == null) {
+            System.out.println("This type doesn't exist.");
+            stmt.setNull(8, 0);
+        } else {
+            stmt.setInt(1, item.getCode());
+            stmt.setString(2, item.getName());
+            stmt.setString(3, item.getShortDesc());
+            stmt.setInt(4, item.getWinrate());
+            stmt.setDate(5, Date.valueOf(item.getReleaseDate()));
+            stmt.setBoolean(6, item.isRanged());
+            stmt.setDouble(7, item.getPrice());
+            stmt.setInt(8, item.getType().getId());
+            affectedRows = stmt.executeUpdate();
+        }
 
         stmt.close();
         return affectedRows;
     }
 
-    public int deleteType(int code) throws SQLException {
-        String query = "DELETE FROM champion WHERE code = ?";
-        PreparedStatement stmt = conn.prepareStatement(query);
+    public int deleteChamp(int code) throws SQLException {
+        int count = 0;
+        if (getChampion(code) == null) {
+            System.out.println("This champion doesn't exist.");
+        } else {
+            String query = "DELETE FROM champion WHERE code = ?";
+            PreparedStatement stmt = conn.prepareStatement(query);
 
-        stmt.setInt(1, code);
+            stmt.setInt(1, code);
 
-        int count = stmt.executeUpdate();
+            count = stmt.executeUpdate();
 
-        stmt.close();
+            stmt.close();
+        }
         return count;
     }
 
-    public int updateType(Champion item) throws SQLException {
-        String query = "UPDATE champion SET name = ?, shortDesc = ?, winrate = ?, releaseDate = ?, isRanged = ?, price = ?, typeID = ? WHERE code = ?";
-        PreparedStatement stmt = conn.prepareStatement(query);
+    public int updateChamp(Champion item) throws SQLException {
+        TypeDAO tpDao = new TypeDAO();
+        int count = 0;
 
-        stmt.setString(1, item.getName());
-        stmt.setString(2, item.getShortDesc());
-        stmt.setInt(3, item.getWinrate());
-        stmt.setDate(4, Date.valueOf(item.getReleaseDate()));
-        stmt.setBoolean(5, item.isRanged());
-        stmt.setDouble(6, item.getPrice());
-        stmt.setInt(7, item.getType().getId());
-        stmt.setInt(8, item.getCode());
+        if (getChampion(item.getCode()) == null) {
+            System.out.println("This champion doesn't exist.");
+        } else if(tpDao.getType(item.getType().getId()) == null){
+            System.out.println("This type doesn't exist.");
+        }
+        else{
+            String query = "UPDATE champion SET name = ?, shortDesc = ?, winrate = ?, releaseDate = ?, isRanged = ?, price = ?, typeID = ? WHERE code = ?";
+            PreparedStatement stmt = conn.prepareStatement(query);
 
-        int count = stmt.executeUpdate();
-        stmt.close();
+            stmt.setString(1, item.getName());
+            stmt.setString(2, item.getShortDesc());
+            stmt.setInt(3, item.getWinrate());
+            stmt.setDate(4, Date.valueOf(item.getReleaseDate()));
+            stmt.setBoolean(5, item.isRanged());
+            stmt.setDouble(6, item.getPrice());
+            stmt.setInt(7, item.getType().getId());
+            stmt.setInt(8, item.getCode());
+
+            count = stmt.executeUpdate();
+            stmt.close();
+        }
         return count;
     }
 
@@ -100,6 +122,7 @@ public class ChampionDAO extends BaseDAO {
         String query = "SELECT * FROM champion ORDER BY code";
         stmt = conn.prepareStatement(query);
         rs = stmt.executeQuery();
+        System.out.println("== ALL CHAMPIONS LIST ==");
         while (rs.next()) {
             System.out.println("================================");
             champ = new Champion();
@@ -115,7 +138,7 @@ public class ChampionDAO extends BaseDAO {
             Type type = tpDao.getType(typeID);
             champ.setType(type);
             champ.showChamp();
-            System.out.println("================================");
+            System.out.println("");
             count++;
         }
 
@@ -132,6 +155,7 @@ public class ChampionDAO extends BaseDAO {
         stmt = conn.prepareStatement(query);
         stmt.setInt(1, filter.getId());
         rs = stmt.executeQuery();
+        System.out.println("== " + filter.getRole() + " champions ==");
         while (rs.next()) {
             System.out.println("================================");
             champ = new Champion();
@@ -147,7 +171,7 @@ public class ChampionDAO extends BaseDAO {
             Type type = tpDao.getType(typeID);
             champ.setType(type);
             champ.showChamp();
-            System.out.println("================================");
+            System.out.println("");
             count++;
         }
         stmt.close();
@@ -167,10 +191,10 @@ public class ChampionDAO extends BaseDAO {
 
         stmt = conn.prepareStatement(query);
         rs = stmt.executeQuery();
-        while(rs.next()){
+        System.out.println("== TOTAL PRICE OF CHAMPIONS PER CHAMPION TYPE ==");
+        while (rs.next()) {
             System.out.println("================================");
             System.out.println(tpDao.getType(rs.getInt("TypeID")).getRole() + ": " + rs.getDouble("sum(price)"));
-            System.out.println("================================");
         }
         stmt.close();
     }
